+++ doc/whatsnew.rst
@@ -22,6 +22,8 @@ v0.7.dev
 
 - Enhance :class:`pyriemann.estimation.Shrinkage` to process HPD matrices. :pr:`307` by :user:`qbarthelemy`
 
+- Add remote sensing examples on radar image clustering. :pr:`306` by :user:`AmmarMian`
+
 v0.6 (April 2024)
 -----------------